requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
python-multipart
jinja2
pydantic-settings
sse-starlette
telethon
